+++ dev/src/DomainLayer/store/Inventory.ts
