+++ dev/src/DomainLayer/user/Authentication.ts
@@ -9,7 +9,6 @@ export class SubscriberData
     public static addSubscriber(subscriber: Subscriber , password:string): void
     {
         var hashedPass : string = createHash('sha1').update(password).digest('hex');
-        console.log(`hash of ${password} is ${hashedPass}`);
         subscriber.setPassword(hashedPass);
         this.subscribers.push(subscriber);
     }
@@ -31,7 +30,6 @@ export class SubscriberData
     public static checkPassword(username: string, password: string) :boolean
     {
         var hashedPass : string = createHash('sha1').update(password).digest('hex');
-        console.log(`hash of ${password} is ${hashedPass}`);
         return this.subscribers.some( user => user.getUsername() === username && user.getPassword() === hashedPass )
     }
