+++ dev/src/DomainLayer/user/User.ts
@@ -1,3 +1,4 @@
+import { Pair } from './ShoppingBasket';
 import { ShoppingCart} from './ShoppingCart'
 
 
@@ -14,6 +15,16 @@ export class User
     public addProductToShoppingCart(storeId: number,  productId: number, quntity: number)
     {
         this.shoppingCart.addProduct(storeId, productId, quntity);
+    }
+
+    public GetShoppingCart(): number[]
+    {
+        return this.shoppingCart.getShoppingCart();
+    }
+
+    public getShoppingBasket(storeId: number): Pair<number, number>[]
+    {
+        return this.shoppingCart.getShoppingBasket(storeId);
     }
