+++ dev/test/store/register.test.ts
